Video-Inpainting/TELEA/requirements.txt
tensorflow
Pillow
numpy
jupyter
matplotlib
PrettyTable
toml
opencv-python==4.3.0.38